Split out mock-generator
Also update json fixtures to use correct date
